[cpp] Changing ms to float

+++ include/Controller.hpp
@@ -1,7 +1,6 @@
 #ifndef INCLUDE_CONTROLLER_HPP
 #define INCLUDE_CONTROLLER_HPP
 
-#include <cstdint>
 class Entity;
 
 
@@ -10,7 +9,7 @@ class Controller
 public:
     virtual ~Controller() = default;
     
-    virtual void apply(Entity*, uint64_t ms) = 0;
+    virtual void apply(Entity*, float ms) = 0;
 
 protected:
     Controller() = default;
