+++ include/World.hpp
@@ -1,7 +1,6 @@
 #ifndef INCLUDE_WORLD_HPP
 #define INCLUDE_WORLD_HPP
 
-#include <cstdint>
 #include <vector>
 #include <memory>
 using std::vector;
@@ -18,7 +17,7 @@ public:
 
     vector<shared_ptr<Entity>> _entities;
 
-    void tick(uint64_t ms);
+    void tick(float ms);
     
 };
